+++ src/lang_utils/error_codes.ml
@@ -181,6 +181,7 @@ let error_codes : (string * string option) list =
     "M0193", None; (* Can't declare actor class to have `async*` result *)
     "M0196", None; (* `system` capability supplied but not required *)
     "M0197", Some([%blob "lang_utils/error_codes/M0197.md"]); (* `system` capability required *)
+    "M0199", Some([%blob "lang_utils/error_codes/M0199.md"]); (* Deprecated experimental stable memory *)
     "M0200", Some([%blob "lang_utils/error_codes/M0200.md"]); (* Cannot determine subtyping or equality *)
     "M0201", None; (* Migration produces/consumes non-stable object *)
     "M0202", None; (* Migration produces/consume non-object type *)
@@ -248,7 +249,6 @@ let warning_codes = [
   "M0194", Some([%blob "lang_utils/error_codes/M0194.md"]), "Unused identifier warning";
   "M0195", Some([%blob "lang_utils/error_codes/M0195.md"]), "warn that `system` capability is implicitly supplied";
   "M0198", Some([%blob "lang_utils/error_codes/M0198.md"]), "Unused field pattern warning";
-  "M0199", Some([%blob "lang_utils/error_codes/M0199.md"]), "Deprecate experimental stable memory"; (* Warn or Error *)
   "M0206", None, "Migration consumes, but does not produce, a declared field";
   "M0207", None, "Migration consumes, but does not produce, an un-declared field";
   "M0210", None, "Parenthetical note must be applied to a message send";