+++ src/mo_config/flags.ml
@@ -86,12 +86,9 @@ let stable_memory_access_limit_default =
   update_call = Int64.mul 1L gigabyte; (* 2 GB limit with 1 GB reserve *)
 }
 let stable_memory_access_limit = ref stable_memory_access_limit_default
-let experimental_stable_memory_default = 0 (* _ < 0: error; _ = 0: warn, _ > 0: allow *)
-let experimental_stable_memory = ref experimental_stable_memory_default
 let typechecker_combine_srcs = ref false (* useful for the language server *)
 let blob_import_placeholders = ref false (* when enabled, blob:file imports resolve as empty blobs *)
 let implicit_derivation_depth = ref 100
-let generate_view_queries = ref false
 
 let default_warning_levels = M.empty
   |> M.add "M0223" Allow (* don't report redundant instantions *)